+++ apache-http-client/src/test/java/com/unidev/http/HttpClientTest.java
@@ -1,18 +1,3 @@
-/*
-  Copyright (c) 2018 Denis O <denis.o@linux.com>
-
-  Licensed under the Apache License, Version 2.0 (the "License");
-  you may not use this file except in compliance with the License.
-  You may obtain a copy of the License at
-
-  http://www.apache.org/licenses/LICENSE-2.0
-
-  Unless required by applicable law or agreed to in writing, software
-  distributed under the License is distributed on an "AS IS" BASIS,
-  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-  See the License for the specific language governing permissions and
-  limitations under the License.
- */
 package com.unidev.http;
 
 import static org.junit.jupiter.api.Assertions.assertNotNull;
