+++ phantom-js/build.gradle
@@ -1,0 +1,4 @@
+dependencies {
+
+    api("com.unidev.platform:unidev-common:4.0.0")
+}
