+++ phantom-js/src/test/java/com/unidev/httpclient/phantomjs/PhantomJSClientTest.java
@@ -1,19 +1,4 @@
-/*
-  Copyright (c) 2018 Denis O <denis.o@linux.com>
-
-  Licensed under the Apache License, Version 2.0 (the "License");
-  you may not use this file except in compliance with the License.
-  You may obtain a copy of the License at
-
-  http://www.apache.org/licenses/LICENSE-2.0
-
-  Unless required by applicable law or agreed to in writing, software
-  distributed under the License is distributed on an "AS IS" BASIS,
-  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-  See the License for the specific language governing permissions and
-  limitations under the License.
- */
-package com.unidev.http;
+package com.unidev.httpclient.phantomjs;
 
 import com.unidev.platform.Processes;
 import com.unidev.httpclient.apache.PhantomJSClient;
